% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Pinpointing on a map where accident happened in a year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{refer to the State Number that we want to plot the accident.}

\item{year}{The year we want to plot the occurence of accidents on the State maps
specified by the state.num parameter.}
}
\value{
A map of the state inputed through the state.num parameter with accidents location
     marked on the map for the desired year.
}
\description{
The function will show us the location on a state map, of where accidents happened
 by the year that we want.
 If an invalid State Number is inputed an error message will be given.
 An error message will also appears if there is no accidents happened in the year
 at the State.
}
\examples{
fars_map_state(1, 2013)
fars_map_state(3, 2015).

}
