% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Year and month the accidents happened.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of years (in integer) that we want to read the data about.}
}
\value{
A data frame with 2 columns specifying the Month and the year for each accidents
   by the years that we put in the list.
}
\description{
The function takes the list of years that we want to know about the accident,
 find the files that record those accidents by the year
 and returning the columns of the months and the year.
 An error "invalid year: <inputed year>" will be printed
 if the data of the year we inputed is not availabe.
}
\examples{
fars_read_years(c(2013, 2014))
fars_read_years(2013)

}
