% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a string 'accident_<specified year>.csv.bz2'}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer that represents the year we want to put in the string.}
}
\value{
A string 'accident_<specified year>.csv.bz2'
}
\description{
By inserting the intended year,
 the function will create a string 'accident_<specified year>.csv.bz2',
 that can be used as a reference to a file name in other functions.
}
\examples{
make_filename(2017)
make_filename(2015)

}
