% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Creating a dataframe table from a csv file.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string that gives the filename of the file that we wish to read.}
}
\value{
A data frame table is returned.
}
\description{
This is a function that first check if the file name inserted exist.
 If it did not, an error is produced ("<filename> does not exist").
 If it did exist the file is read, and the data is transformed
 into a table of data frame.
}
\examples{
# Setting the path for the example data
expath <- system.file("extdata", package="farsForCoursera")
setwd(expath)
fars_read("accident_2013.csv.bz2")

}
